Extract tool spec helpers into codex-tools
