+++ codex-rs/tools/BUILD.bazel
@@ -3,4 +3,7 @@ load("//:defs.bzl", "codex_rust_crate")
 codex_rust_crate(
     name = "tools",
     crate_name = "codex_tools",
+    compile_data = [
+        "src/tool_apply_patch.lark",
+    ],
 )
