+++ codex-rs/tools/src/lib.rs
@@ -3,6 +3,7 @@
 
 mod agent_job_tool;
 mod agent_tool;
+mod apply_patch_tool;
 mod code_mode;
 mod dynamic_tool;
 mod image_detail;
@@ -35,6 +36,9 @@ pub use agent_tool::create_spawn_agent_tool_v1;
 pub use agent_tool::create_spawn_agent_tool_v2;
 pub use agent_tool::create_wait_agent_tool_v1;
 pub use agent_tool::create_wait_agent_tool_v2;
+pub use apply_patch_tool::ApplyPatchToolArgs;
+pub use apply_patch_tool::create_apply_patch_freeform_tool;
+pub use apply_patch_tool::create_apply_patch_json_tool;
 pub use code_mode::augment_tool_spec_for_code_mode;
 pub use code_mode::create_code_mode_tool;
 pub use code_mode::create_wait_tool;
@@ -54,12 +58,15 @@ pub use local_tool::create_request_permissions_tool;
 pub use local_tool::create_shell_command_tool;
 pub use local_tool::create_shell_tool;
 pub use local_tool::create_write_stdin_tool;
+pub use local_tool::request_permissions_tool_description;
 pub use mcp_resource_tool::create_list_mcp_resource_templates_tool;
 pub use mcp_resource_tool::create_list_mcp_resources_tool;
 pub use mcp_resource_tool::create_read_mcp_resource_tool;
 pub use mcp_tool::mcp_call_tool_result_output_schema;
 pub use mcp_tool::parse_mcp_tool;
 pub use request_user_input_tool::create_request_user_input_tool;
+pub use request_user_input_tool::request_user_input_tool_description;
+pub use request_user_input_tool::request_user_input_unavailable_message;
 pub use responses_api::FreeformTool;
 pub use responses_api::FreeformToolFormat;
 pub use responses_api::ResponsesApiNamespace;
