+++ codex-rs/tools/src/request_user_input_tool_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use codex_protocol::config_types::ModeKind;
 use pretty_assertions::assert_eq;
 use std::collections::BTreeMap;
 
@@ -98,5 +99,56 @@ fn request_user_input_tool_includes_questions_schema() {
             },
             output_schema: None,
         })
+    );
+}
+
+#[test]
+fn request_user_input_unavailable_messages_respect_default_mode_feature_flag() {
+    assert_eq!(
+        request_user_input_unavailable_message(
+            ModeKind::Plan,
+            /*default_mode_request_user_input*/ false
+        ),
+        None
+    );
+    assert_eq!(
+        request_user_input_unavailable_message(
+            ModeKind::Default,
+            /*default_mode_request_user_input*/ false
+        ),
+        Some("request_user_input is unavailable in Default mode".to_string())
+    );
+    assert_eq!(
+        request_user_input_unavailable_message(
+            ModeKind::Default,
+            /*default_mode_request_user_input*/ true
+        ),
+        None
+    );
+    assert_eq!(
+        request_user_input_unavailable_message(
+            ModeKind::Execute,
+            /*default_mode_request_user_input*/ false
+        ),
+        Some("request_user_input is unavailable in Execute mode".to_string())
+    );
+    assert_eq!(
+        request_user_input_unavailable_message(
+            ModeKind::PairProgramming,
+            /*default_mode_request_user_input*/ false
+        ),
+        Some("request_user_input is unavailable in Pair Programming mode".to_string())
+    );
+}
+
+#[test]
+fn request_user_input_tool_description_mentions_available_modes() {
+    assert_eq!(
+        request_user_input_tool_description(/*default_mode_request_user_input*/ false),
+        "Request user input for one to three short questions and wait for the response. This tool is only available in Plan mode.".to_string()
+    );
+    assert_eq!(
+        request_user_input_tool_description(/*default_mode_request_user_input*/ true),
+        "Request user input for one to three short questions and wait for the response. This tool is only available in Default or Plan mode.".to_string()
     );
 }
